Use text-decoration-thickness value in underline painting
Parsing of text-decoration-thickness was added in [1], use the parsed
value in text decoration thickness painting.

Add a WPT ref test that compares text-decoration-thickness value from a
variable font (keyword value 'from-font' for text-decoration-thickness)
with the identical rendering achieved by using a static font with thick
and thin underline metadata.

Add a second WPT ref test that compares a fixed
text-decoration-thickness value to the from-font values of a font with
thick and thin underlines.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2178888

Bug: 785230
Change-Id: I626a321a38c77e106215b640dcf85c20f7a0ad93